+++ src/components/CardRow.tsx
@@ -1,6 +1,7 @@
-import { useEffect, useRef } from 'react';
+import { useRef } from 'react';
 import { getContentText, getContentImageURL, getContentIDs } from '../types/helpers.ts';
 import { type CardRowProps, ContentFields } from '../types/types.ts';
+import { horizontalScroll } from '../hooks/horizontalScroll.ts';
 import '../styles/CardRow.css'
 
 const PLACEHOLDER_IMAGE = "/disney-plus-placeholder.png";
@@ -8,44 +9,23 @@ const PLACEHOLDER_IMAGE = "/disney-plus-placeholder.png";
 // React Component that handles the rendering of a given row 
 
 export function CardRow({title, items, loading, error, selectedIndex}:CardRowProps) {
-    const containerRef = useRef<HTMLDivElement>(null);
-    const cardRefs = useRef<(HTMLDivElement | null)[]>([]);
+  const containerRef = useRef<HTMLDivElement>(null);
+  const cardRefs = useRef<(HTMLDivElement | null)[]>([]);
 
-    const cards = items.map(item => ({
-        id: getContentIDs(item),
-        title: getContentText(item, ContentFields.TEXT_FULL, ContentFields.TEXT_TITLE),
-        tile: getContentImageURL(item, ContentFields.IMAGE_TILE, ContentFields.IMAGE_TILE_RATIO),
-    }));
+  const cards = items.map(item => ({
+      id: getContentIDs(item),
+      title: getContentText(item, ContentFields.TEXT_FULL, ContentFields.TEXT_TITLE),
+      tile: getContentImageURL(item, ContentFields.IMAGE_TILE, ContentFields.IMAGE_TILE_RATIO),
+  }));
 
-    useEffect(() => {
-      if (selectedIndex >= 0 && cardRefs.current[selectedIndex] && containerRef.current) {
-          const selectedCard = cardRefs.current[selectedIndex];
-          const container = containerRef.current;
-
-          if (selectedCard && container) {
-              const cardRect = selectedCard.getBoundingClientRect();
-              const containerRect = container.getBoundingClientRect();
-
-              if (cardRect.left < containerRect.left || cardRect.right > containerRect.right) {
-                  const cardCenter = selectedCard.offsetLeft + selectedCard.offsetWidth / 2;
-                  const containerCenter = container.offsetWidth / 2;
-                  const scrollPosition = cardCenter - containerCenter;
-
-                  container.scrollTo({
-                      left: Math.max(0, scrollPosition),
-                      behavior: "smooth"
-                  });
-              }
-          }
-      }
-  }, [selectedIndex]);
-  
   const handleImgError = (element: React.SyntheticEvent<HTMLImageElement, Event>) => {
       const img = element.currentTarget;
       if(img.src !== PLACEHOLDER_IMAGE) {
         img.src = PLACEHOLDER_IMAGE;
       }
   };
+
+  horizontalScroll(selectedIndex, containerRef, cardRefs);
 
   if(loading) return <div>Loading Disney+ {title} stuff...</div>;
   if(error) return <div>Error: {error}</div>;
